Use libdimentio static library + put extra functions on their own
